+++ src/commcare_cloud/commands/terraform/aws.py
@@ -1,7 +1,6 @@
 # coding=utf-8
 from __future__ import absolute_import, print_function, unicode_literals
 
-import getpass
 import json
 import os
 import subprocess
@@ -14,7 +13,6 @@ from io import open
 import boto3
 import jinja2
 import requests
-import six
 import yaml
 from clint.textui import puts
 from memoized import memoized
@@ -24,6 +22,7 @@ from six.moves import configparser, input, shlex_quote
 from commcare_cloud.cli_utils import print_command
 from commcare_cloud.colors import color_notice, color_success
 from commcare_cloud.commands.command_base import Argument, CommandBase
+from commcare_cloud.user_utils import get_default_username
 from commcare_cloud.environment.main import get_environment
 
 
@@ -290,29 +289,6 @@ class AwsFillInventoryHelper(object):
 
 
 DEFAULT_SIGN_IN_DURATION_MINUTES = 30
-
-
-class StringIsGuess(six.text_type):
-    def __new__(cls, *args, **kwargs):
-        is_guess = kwargs.pop('is_guess')
-        self = super(StringIsGuess, cls).__new__(cls, *args, **kwargs)
-        self.is_guess = is_guess
-        return self
-
-
-@memoized
-def get_default_username():
-    """
-    Returns a special string type that has field is_guess
-
-    If is_guess is True, the caller should assume the user wants this value
-    and should not give them a chance to change their choice of user interactively.
-    """
-    environ_username = os.environ.get('COMMCARE_CLOUD_DEFAULT_USERNAME')
-    if environ_username:
-        return StringIsGuess(environ_username, is_guess=False)
-    else:
-        return StringIsGuess(getpass.getuser(), is_guess=True)
 
 
 def print_help_message_about_the_commcare_cloud_default_username_env_var(username):
